Added dialog to confirm deletion of food entries

+++ Software/food_expiration_reminder/lib/src/settings.dart
@@ -9,16 +9,53 @@ class Settings extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        body: Center(
-          child: TextButton(
-            child: Text('Reset'),
-            onPressed: deleteAllFoodEntries,
-            ),
+      body: Center(
+        child: TextButton(
+          child:
+              Text('Delete all food reminders', 
+              style: TextStyle(fontSize: 20)),
+          onPressed: () => _showMyDialog(context),
         ),
+      ),
     );
   }
 
-    void deleteAllFoodEntries() {
-      storage.deleteFile();
+  void deleteAllFoodEntries() {
+    storage.deleteFile();
+  }
+
+  Future<void> _showMyDialog(BuildContext context) async {
+    return showDialog<void>(
+      context: context,
+      barrierDismissible: false, // user must tap button!
+      builder: (BuildContext context) {
+        return AlertDialog(
+          title: Text('Reset reminders'),
+          content: SingleChildScrollView(
+            child: ListBody(
+              children: <Widget>[
+                Text('Are you sure you want to delete all the reminders?'),
+                Text('This action is irreversible.'),
+              ],
+            ),
+          ),
+          actions: <Widget>[
+            TextButton(
+              child: Text('Yes'),
+              onPressed: () {
+                deleteAllFoodEntries();
+                Navigator.of(context).pop();
+              },
+            ),
+            TextButton(
+              child: Text('No'),
+              onPressed: () {
+                Navigator.of(context).pop();
+              },
+            ),
+          ],
+        );
+      },
+    );
   }
 }
